requirements.txt
BareNecessities==0.2.8
blinker==1.4
cachelib==0.7.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Mail==0.9.1
Flask-Session==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.1
Mail==2.1.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==3.20.1
validate-email==1.3
Werkzeug==2.1.1
